requirements.txt
click==8.1.7
openai-whisper==20231117